feat(): Implement the begin of nested routing
What the label says.
Add a way to add links to the tab components which will be doing the route changes.
Listen for clicks on tabs and map that to the provided links.
The links are generated using router.createHref() so that the component does not
have to know what route it is 'mounted' at. This is why the routing parts pass the
router.path(path) stuff :smile: I was planning ahead.